DemoPlugin update version 6.2
- Restructured project to use src/, Actions/, Helpers/, images/ folders
- Added ToggleMuteCommand with macOS and Windows support
- Added ButtonSwitchesCommand with 4 toggle switches
- Added ThumbUpDownCommand with embedded image toggle
- Added CounterAdjustment for dial rotation tracking
- Added .vscode config for debugging and build tasks
- Added README.md for new developers
