+++ DemoPlugin/.vscode/tasks.json
@@ -1,0 +1,40 @@
+{
+  "version": "2.0.0",
+  "tasks": [
+    {
+      "label": "Build (Debug)",
+      "type": "shell",
+      "command": "dotnet build -c Debug",
+      "group": "build",
+      "problemMatcher": "$msCompile"
+    },
+    {
+      "label": "Build (Release)",
+      "type": "shell",
+      "command": "dotnet build -c Release",
+      "group": "build",
+      "problemMatcher": "$msCompile"
+    },
+    {
+      "label": "Package Plugin",
+      "type": "shell",
+      "command": "logiplugintool pack ./bin/Release ./Demo.lplug4",
+      "group": "none",
+      "problemMatcher": []
+    },
+    {
+      "label": "Install Plugin",
+      "type": "shell",
+      "command": "logiplugintool install ./Demo.lplug4",
+      "group": "none",
+      "problemMatcher": []
+    },
+    {
+      "label": "Uninstall Plugin",
+      "type": "shell",
+      "command": "logiplugintool uninstall Demo",
+      "group": "none",
+      "problemMatcher": []
+    }
+  ]
+}
