--- DemoPlugin/DemoPlugin/DemoPlugin.csproj.user
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup Condition=" $(OS) == 'Windows_NT' ">
-    <StartAction>Program</StartAction>
-    <StartProgram>C:\Program Files\Logi\LogiPluginService\LogiPluginService.exe</StartProgram>
-  </PropertyGroup>
-  <PropertyGroup Condition=" $(OS) != 'Windows_NT' and '$(RunConfiguration)' == 'Default' ">
-    <StartAction>Program</StartAction>
-    <StartProgram>/Applications/Utilities/LogiPluginService.app</StartProgram>
-  </PropertyGroup>
-</Project>
