--- DemoPlugin/DemoPlugin/package/icontemplates/Loupedeck.DemoPlugin.ToggleMuteCommand.ict
@@ -1,37 +1,0 @@
-{
-  "backgroundColor": 4278190080,
-  "items": [
-    {
-      "$type": "Loupedeck.Service.ActionIconImageItem, LoupedeckShared",
-      "image": "iVBORw0KGgoAAAANSUhEUgAAAFAAAABQCAYAAACOEfKtAAAABGdBTUEAALGPC/xhBQAAAAlwSFlzAAABwwAAAcMBOnufsAAAABh0RVh0U29mdHdhcmUAcGFpbnQubmV0IDQuMS42/U4J6AAAAe5JREFUeF7t2rFNA0EQhWEHFEIhBJRCCZRARoioAHLojgKOHYmVVqvf5glm70bWG+lL/l07mMA+LE7bttk/YDQdRtNhNB1G02E0HUbTYTQdRtNhNB1G02E0HUbTYTQdRtNhNB1G02E0HUbTYTQdRtNhNB1G02E0HUbTYTQdxkLeoZWCsYinJobOysBYwEPTh87LwHiwcXkxdKcMjAealxdD98rAeBBaXgzdLQPjAc4tL4bul4FxZ5eWF0OvKQNjstvm/ozH5reh97xrnqc2irO4Q2epMCbrz3N/nfn9YjFfTcxLM59Hi4k7y5eIMVn2At+accYl9uX1ibvja9NhTJa9wJvmoxknFjcvL+7E3fn1qTAmy15goCWOs8vyAsZkKxYYzi1xt+UFjMlWLTC8NvNEo7tLYEy2aoHzZ9444xfLUhiTrVjgvLzPH+PsskSMybIXGA/J4/TPPPpMvPSwnQJjsuwFjg/S8xfGuMSreZBe9adcPCTTt220OFu+vIBxZ/4xIYF/zkrgH1QT0BLpXhkYDzYvke6UgbGAcYl0XgbGIvrzI52VgbEQ/2vHtcNoOoymw2g6jKbDaDqMpsNoOoymw2g6jKbDaDqMpsNoOoymw2g6jKbDaDqMpsNoOoymw2g6jKbDaDqMpsNoqu30DbZ472o00s59AAAAAElFTkSuQmCC",
-      "imageFileName": "ToggleMute.png",
-      "imageColor": 4294967295,
-      "imageRotation": "None",
-      "isVisible": true,
-      "itemType": "Image",
-      "area": {
-        "x": 0,
-        "y": -10,
-        "width": 100,
-        "height": 100,
-        "isFullScreen": true
-      }
-    },
-    {
-      "$type": "Loupedeck.Service.ActionIconTextItem, LoupedeckShared",
-      "text": "Toggle Mute",
-      "textColor": 4294967295,
-      "fontSize": 5,
-      "fontName": "Brown Logitech Pan Light",
-      "isVisible": true,
-      "itemType": "Text",
-      "area": {
-        "x": 0,
-        "y": 70,
-        "width": 100,
-        "height": 16,
-        "isFullScreen": false
-      }
-    }
-  ]
-}
