--- DemoPlugin/DemoPlugin/package/metadata/DefaultIconTemplate.ict
@@ -1,33 +1,0 @@
-{
-  "backgroundColor": 4278869247,
-  "items": [
-    {
-      "image": "",
-      "imageFileName": null,
-      "imageColor": 4294967295,
-      "imageRotation": "None",
-      "isVisible": true,
-      "itemType": "Image",
-      "area": {
-        "x": 15,
-        "y": 0,
-        "width": 70,
-        "height": 70
-      }
-    },
-    {
-      "text": "",
-      "textColor": 4294967295,
-      "fontSize": 5,
-      "fontName": "Brown Logitech Pan Light",
-      "isVisible": true,
-      "itemType": "Text",
-      "area": {
-        "x": 0,
-        "y": 70,
-        "width": 100,
-        "height": 30
-      }
-    }
-  ]
-}
