--- DemoPlugin/DemoPlugin/package/metadata/LoupedeckPackage.yaml
@@ -1,53 +1,0 @@
-# ==================================================================================================
-# GENERAL SETTINGS
-# ==================================================================================================
-
-# Package type. Must be plugin4 for plugins.
-type: plugin4
-
-# Name that identifies the plugin.
-name: Demo
-
-# Plugin name that is shown to the user.
-displayName: Demo
-
-# Short description of the plugin.
-description: This is a demo plugin
-
-# Name of the plugin library file.
-pluginFileName: DemoPlugin.dll
-
-# Plugin version.
-version: 1.0
-
-# Author of the plugin. The author can be a company or an individual developer.
-author: Logitech Europe SA
-
-# Copyright information.
-copyright: Copyright © 2025 Logitech Europe SA. All rights reserved.
-
-# Minimum SDK version supported by the plugin.
-minimumLoupedeckVersion: 6.0
-
-
-# ==================================================================================================
-# PLUGIN PROPERTIES
-# ==================================================================================================
-
-# Location of plugin files on Windows (relative to the plugin base directory).
-# This parameter is required to support Windows.
-pluginFolderWin: bin
-
-# Location of plugin files on macOS (relative to the plugin base directory).
-# This parameter is required to support Mac.
-pluginFolderMac: bin
-
-# List of devices the plugin is optimized for.
-supportedDevices:
-    # LoupedeckCtFamily covers the following devices: Loupedeck CT, Live and Live S,
-    # Razer Stream Controller, and Razer Stream Controller X.
-    - LoupedeckCtFamily
-
-    # Logitech MX Creative Console family devices.
-    - LoupedeckExtendedFamily
-
