+++ DemoPlugin/src/DemoApplication.cs
@@ -2,6 +2,8 @@ namespace Loupedeck.DemoPlugin
 {
     using System;
 
+    // This class can be used to connect the Loupedeck plugin to an application.
+
     public class DemoApplication : ClientApplication
     {
         public DemoApplication()
@@ -9,9 +11,12 @@ namespace Loupedeck.DemoPlugin
         }
 
         // This method can be used to link the plugin to a Windows application.
-        protected override String GetProcessName() => "DemoApplication";
+        protected override String GetProcessName() => "";
 
         // This method can be used to link the plugin to a macOS application.
         protected override String GetBundleName() => "";
+
+        // This method can be used to check whether the application is installed or not.
+        public override ClientApplicationStatus GetApplicationStatus() => ClientApplicationStatus.Unknown;
     }
 }
