+++ DemoPlugin/src/Directory.Build.props
@@ -2,6 +2,6 @@
     <PropertyGroup>
         <LangVersion>latest</LangVersion>
         <BaseIntermediateOutputPath>$(SolutionDir)obj\</BaseIntermediateOutputPath>
-        <BaseOutputPath>$(SolutionDir)bin\</BaseOutputPath>
+        <BaseOutputPath>$(SolutionDir)..\bin\</BaseOutputPath>
     </PropertyGroup>
 </Project>
