+++ DemoPlugin/src/Helpers/PluginResources.cs
