--- README.md
@@ -1,12 +1,0 @@
-# Welcome to Logi Actions SDK
-
-Here you can find an introduction to Logi Actions SDK, example plugin source code and the SDK documentation. 
-
-## Documentation
-
-* [Logi Actions SDK Developer Docs](https://logitech.github.io/actions-sdk-docs)
-
-## Developer Discord
-
-* [Logi Developer Discord](https://discord.gg/etJCPZytHg)
-
